refactor: drop the repoquery feature and rename it to who_needs
The `repoquery` feature became vestigial once the pure matcher was made
`pub(crate)`: building `--features repoquery` alone compiled, but every
item in the module was dead code, since the gateway is the only consumer.
The `gateway` feature already pulled `rattler_conda_types` in via
`sparse`, so removing the feature costs nothing; the module is now gated
on `gateway` instead.

"repoquery" was also a poor name for what the module holds. Renamed
throughout, keeping the two-file split that separates *what* counts as a
reverse dependency from *where* the records come from:

- `src/repoquery.rs` -> `src/who_needs.rs` (types + pure matcher, no I/O)
- `src/gateway/who_needs.rs` -> `src/gateway/who_needs_query.rs`
  (`WhoNeedsQuery`: source resolution, batching, concurrency)

The `_query` suffix avoids two files named `who_needs.rs` in one crate.
Type names are unchanged. The Python side follows suit
(`rattler.repo_data.repoquery` -> `rattler.repo_data.who_needs`); both
modules are new in this PR, so nothing published depends on the old
names.
