+++ crates/rattler_repodata_gateway/Cargo.toml
@@ -114,8 +114,7 @@ default = ['rustls']
 native-tls = ['reqwest/native-tls', 'rattler_networking/native-tls', 'rattler_cache/native-tls', 'rattler_redaction/native-tls']
 rustls = ['reqwest/rustls', 'rattler_networking/rustls', 'rattler_cache/rustls', 'rattler_redaction/rustls']
 sparse = ["rattler_conda_types", "memmap2", "self_cell", "superslice", "itertools", "serde_json/raw_value"]
-repoquery = ["rattler_conda_types"]
-gateway = ["sparse", "repoquery", "http", "http-cache-semantics", "parking_lot", "async-trait", "rattler_package_streaming"]
+gateway = ["sparse", "http", "http-cache-semantics", "parking_lot", "async-trait", "rattler_package_streaming"]
 
 [package.metadata.docs.rs]
 features = ["sparse", "gateway"]
