--- py-rattler/docs/repoquery.md
@@ -1,3 +1,0 @@
-# Repoquery
-
-::: rattler.repo_data.repoquery
