+++ py-rattler/docs/who_needs.md
@@ -1,0 +1,3 @@
+# Repoquery
+
+::: rattler.repo_data.who_needs
